server player extends connection
